More work on notifications
