+++ code/app/src/main/java/com/example/booknet/BookListing.java
@@ -1,10 +1,7 @@
 package com.example.booknet;
 
-import android.util.Log;
-import android.graphics.Bitmap;
 
 import java.io.Serializable;
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 /**
@@ -12,36 +9,9 @@ import java.util.ArrayList;
  */
 public class BookListing implements Serializable, Cloneable {
 
-    /**
-     * Enum for the status of a BookListing, so the values are more easily tracked
-     */
-    public enum Status {
-        Available, Requested, Accepted, Borrowed;
-
-        /**
-         * Returns the status as a text string.
-         *
-         * @return
-         */
-        @Override
-        public String toString() {
-            switch (this) {
-                case Available:
-                    return "Available";
-                case Requested:
-                    return "Requested";
-                case Accepted:
-                    return "Accepted";
-                case Borrowed:
-                    return "Borrowed";
-            }
-            return super.toString();
-        }
-    }
-
     //Attributes
     private Book book;
-    private Status status;
+    private BookListingStatus status;
     private String ownerUsername;
     private ArrayList<String> requests;
     private String borrowerName;
@@ -54,7 +24,7 @@ public class BookListing implements Serializable, Cloneable {
         this.book = new Book();
         this.ownerUsername = "";
         this.borrowerName = "";
-        this.status = Status.Available;
+        this.status = BookListingStatus.Available;
         this.requests = new ArrayList<String>();
         this.geoLocation = new UserLocation();
 
@@ -69,7 +39,7 @@ public class BookListing implements Serializable, Cloneable {
         this.book = book;
         this.ownerUsername = CurrentUser.getInstance().getUsername();
         this.borrowerName = "";
-        this.status = Status.Available;
+        this.status = BookListingStatus.Available;
         this.requests = new ArrayList<String>();
         this.geoLocation = new UserLocation();
     }
@@ -79,9 +49,7 @@ public class BookListing implements Serializable, Cloneable {
         return book;
     }
 
-    public Status getStatus() {
-        return status;
-    }
+    public BookListingStatus getStatus() { return status; }
 
     public String getOwnerUsername() {
         return ownerUsername;
@@ -114,13 +82,13 @@ public class BookListing implements Serializable, Cloneable {
      */
     public void addRequest(String requesterName) {
         //Only requestable when Available or Requested
-        if (status == Status.Available || status == Status.Requested) {
+        if (status == BookListingStatus.Available || status == BookListingStatus.Requested) {
             //Add to the list
             if (!requests.contains(requesterName)) {
                 requests.add(requesterName);
             }
             //change status to reflect change
-            this.status = Status.Requested;
+            this.status = BookListingStatus.Requested;
             //todo send to database
         } else {
             //todo notify cannot add request
@@ -133,7 +101,7 @@ public class BookListing implements Serializable, Cloneable {
      * @param requesterName The user whose request to accept
      */
     public void acceptRequest(String requesterName) {
-        if (status == Status.Requested) {
+        if (status == BookListingStatus.Requested) {
             if (requests.contains(requesterName)) {
                 requests.remove(requesterName);
 
@@ -146,7 +114,7 @@ public class BookListing implements Serializable, Cloneable {
                 requests.clear();
 
                 //Accept this request
-                status = Status.Accepted;
+                status = BookListingStatus.Accepted;
                 this.borrowerName = requesterName;
                 //todo allow geolocation
                 //todo notify database
@@ -174,7 +142,7 @@ public class BookListing implements Serializable, Cloneable {
     public void denyRequest(String requesterName) {
         if (requests.contains(requesterName)) {
             requests.remove(requesterName);
-            if(requests.size() < 1) this.status = Status.Available;
+            if(requests.size() < 1) this.status = BookListingStatus.Available;
             //todo notify database
         }
     }
@@ -184,7 +152,7 @@ public class BookListing implements Serializable, Cloneable {
      */
     public void bookBorrowed() {
         //todo: complete?
-        status = Status.Borrowed;
+        status = BookListingStatus.Borrowed;
     }
 
     /**
@@ -192,7 +160,7 @@ public class BookListing implements Serializable, Cloneable {
      */
     public void bookReturned() {
         //todo: complete?
-        status = Status.Available;
+        status = BookListingStatus.Available;
         borrowerName = "";
     }
 
@@ -235,7 +203,7 @@ public class BookListing implements Serializable, Cloneable {
         this.requests = requests;
     }
 
-    public void setStatus(Status status) {
+    public void setStatus(BookListingStatus status) {
         this.status = status;
     }
