--- code/app/src/main/java/com/example/booknet/NotificationActivity.java
@@ -1,4 +1,0 @@
-package com.example.booknet;
-
-public class NotificationActivity {
-}
