+++ code/app/src/main/java/com/example/booknet/NotificationFragment.java
@@ -1,0 +1,6 @@
+package com.example.booknet;
+
+import android.support.v4.app.Fragment;
+
+public class NotificationFragment extends Fragment {
+}
